+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ### Unreleased
 
+### 0.9.0
+
 * Dedent current line on `else`, `elif`, `except`, and `finally` statements
     * ***NOTE:*** You no longer need to manually un-indent on these lines, just press `Enter` and it will be un-indented for you.
 * Don't dedent on keywords if they appear in triple quoted string.
+* Update demo gif to showcase improved dedent behaviors.
 
 ### 0.8.1